==> scripts/ESR 2026/01_create_networks_cciea_ESR2026.Rmd <==
---
title: "Create fisheries participation networks for CCIEA Ecosystem Status Report to the Pacific Fishery Management Council"
author: "Jameal Samhouri, Mary Fisher"
date: "Written 12/29/2021. Last Run `r Sys.Date()`"
output: 
  html_document:
    df_print: paged
    toc: yes
    toc_depth: '5'
  pdf_document:
    highlight: haddock
    number_sections: yes
    toc: yes
    toc_depth: '3'
geometry: margin=1in
fontsize: 11pt
---

# Description

This document generates fisheries participation networks for US West Coast port groups using the following approach:

- networks are based on fisheries participation and revenues at the vessel level, rather than aggregated across vessels for individual ports or states;
- edges in the networks are based on a connectivity definition (the amount and evenness of revenue across pairs of nodes);
- nodes (fisheries) that do not contribute to at least a median of 10% of the annual revenue of associated vessels are excluded;
- nodes (fisheries) are included in each network with a minimum total fisheries revenue cutoff for vessels of `$5,000` (USD in 2005; following Holland et al. diversification indices) and a minimum revenue cutoff for individual fisheries of `$500` (USD in 2005; following Holland et al. diversification indices)

The code makes network graphs and calculates network statistics.

A related file "02_plot_annual_networks_cciea.Rmd" visually summarizes a subset of these networks for presentation.

# Setup

Clear the environment
```{r}
rm(list=ls())
```

<br>
```{r "setup", include=FALSE}
if(!require("here")) {install.packages("here")}
library(here)
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = here::here())

## start time for full script
script_start_time <- Sys.time()
```
<br>

This script requires the following packages. 
```{r packages, message=FALSE, warning=FALSE}
if(!require("tidyverse")) {install.packages("tidyverse")}
if(!require("lubridate")) {install.packages("lubridate")}
if(!require("ggplot2")) {install.packages("ggplot2")}
if(!require("igraph")) {install.packages("igraph")}
```
<br>

And calls the following functions:
```{r}
source("R/generate_participation_network.R") # this function is for vessel-level networks, summarized at the port or state scale
source("R/color_network_nodes.R")
source("R/network_stats_functions.R")
source("R/calc_network_stats.R")
```
<br>

# User Inputs 

Select your directories.
```{r}
## location of fish ticket with metiers assigned, metier key for each port group
indir = '/Users/jameal.samhouri/Documents/Github repos/CCIEA Networks/processed'
# indir = "/Users/abigail.golden/Desktop/Research/24 - CEFI/data/fish_tickets/processed" # Abigail's fish ticket directory

## output directory for igraph objects
outdir_ports = 'data/networks/participation_vessel_ports'
# outdir_states = 'data/networks/participation_vessel_states'

## output directory for network viz
pngdir_ports = 'data/networks/participation_vessel_ports/plots'
# pngdir_states = 'data/networks/participation_vessel_states/plots'

## output file (including directory) for stat summaries
statdir = 'results/statistics'

## directory for species group key. not confidential
processdir <- "data/input"
```
<br>

Specify file name of the species groupings lookup key that contains labels for the network graphs.
```{r get_filenames}

myfile5 <- "spgrpn2_iea_key.csv"

```
<br>

Specify file name for inflation adjustment factors.
```{r get_AFIvalues}

gdp_defl <- read.csv(file = paste(here(processdir), "inflation_adjustment.csv", sep = "/")) # generated in script 00_process

```


Identify the crab years, port groups, and states that you would like to produce networks for. The port groups vector should include only those port groups which are present in the single data file produced with script 00. The states vector should include only those states which are present in the agid column of the fishtix data produced with script 00. 
```{r}
## (crab) years
years <- seq(2004,2025)

## IOPAC port groups
myports <- c("Puget Sound","North WA Coast","WA Coast","Other Coastal Washington","Astoria","Tillamook","Columbia River","Newport","Coos Bay","Brookings","Crescent City","Eureka","Morro Bay","Fort Bragg","Bodega Bay","San Francisco","San Diego","Monterey","Santa Barbara","Los Angeles","Unknown Ports" )

# wa_ports <- myports[1:4]
# or_ports <- myports[5:10]
# ca_ports <- myports[11:20]

## west coast states
mystates <- c("C", "O","W")

```
<br>

Some Dungeness crab landings may be recorded prior to the official opening date of the season as a part of domoic acid testing. We remove these landings because we are interested in flows of fishers between fisheries as a result of within-season activity.
```{r}
rm_crab = TRUE
```
<br>

For confidentiality, three or more vessels must be participating in a given fishery for that fishery to be included in the networks. To avoid inclusion of rare or minimal fishing activity, a given fishery must contribute to at least 10% of a vessel's seasonal revenue to be included in the network data. 

The cutoff values can be increased if more stringent fishery inclusion criteria are preferred, and the contribution cutoff may be decreased (i.e. for summaries over shorter temporal / smaller spatial scales).
```{r}

vessel_cutoff <- 3
contr_cutoff <- 0.10

```
<br>


Set these objects according to your user inputs from Scripts 1 and 3
```{r}
## the value of `k` for the k-nearest neighbor function
k <- 1

## the reference years first run through infomap
## for the cciea network analysis this does not matter because we use a prior sp groupings
ref_years <- c(2005,2006)
```
<br>


# Generate all networks

Develop vessel-level networks summarized by port group (n=21).

We will make these network graphs for crab years 2005-2025. 


## Read in data

Read in the landings data, from the file containing all fish tickets across port groups, within a single crab year.

NOTE: this step takes a while.

NOTES FOR JAMEAL: 
1) read in fish ticket files with this naming convention: fish_tickets_crab2004_processed_for_networks_2024-12-31.csv
2) if species groupings column is named something other than metier.name, you'll have to go through the generate_participation_network.R function to replace it (done, JS note 010521)
```{r}

# date of last fish ticket processing
# rundate <- "2025-01-08" # Abigail's last fish ticket processing
rundate <- "2025-12-23" # Jameal's last fish ticket processing

for(y in years){
  if(y %in% ref_years){
    tmptix <- read.csv(paste0(indir,"/fish_tickets_crab", y, "_processed_for_networks_", rundate, ".csv"), stringsAsFactors = FALSE) %>%
    filter(IOPAC %in% myports)
  } else{
    tmptix <- read.csv(paste0(indir,"/fish_tickets_crab", y, "_processed_for_networks_" , rundate, ".csv"), stringsAsFactors = FALSE) %>%
    filter(IOPAC %in% myports)
  }
  if(exists('fishtix')){
    fishtix <- rbind(fishtix, tmptix)
  } else{ fishtix <- tmptix }
}
rm(tmptix)

# some checks on species groups
unique(fishtix$SPGRPN2)
length(which(is.na(fishtix$SPGRPN2)))
unique(fishtix[which(is.na(fishtix$SPGRPN2)),]$spid)
unique(fishtix[which(is.na(fishtix$SPGRPN2)),]$year)
unique(fishtix[which(is.na(fishtix$SPGRPN2)),]$agid)
unique(fishtix[which(is.na(fishtix$SPGRPN2)),]$pcid)

min(fishtix$year)
max(fishtix$year)
min(fishtix$crab_year)
max(fishtix$crab_year)

#View(fishtix %>% filter(crab_year == 2004))

```
<br>

Make sure that all dates are `Posixt` objects and species group labels are characters.
```{r}
#dates_df$odate <- mdy(dates_df$odate)
fishtix$tdate <- date(parse_date_time(fishtix$tdate, orders=c("ymd", "mdy")))
fishtix$SPGRPN2 = as.character(fishtix$SPGRPN2)
```
<br>

Node names
```{r}

#### Read in species group labels key ####
node_names <- read_csv(here::here(processdir,myfile5),
                       col_types = list(col_character(), col_character(), col_character()))

```
<br>

## Make Participation Networks

For each crab year and port group, this code will:
1. Create a `graph` object for each fisheries participation network, and then save the object as an R data file

2. Plot the network graphs

3. Calculate network-level statistics, and save to an output data frame.

### 1) Vessel-level networks summarized by port group

We only include vessels that generate >=`$5,000` (USD in 2005; following Holland et al. diversification indices) in total fisheries revenue each year and >=`$500` (USD in 2005; following Holland et al. diversification indices) in revenue from individual fisheries.

```{r}

total_rev_cutoff_2005 <- 5000 #USD in 2005; following Holland et al. diversification indices
indiv_rev_cutoff_2005 <- 500 #USD in 2005; following Holland et al. diversification indices

gdp_defl_2005 <- gdp_defl %>% filter ( year == 2005) %>% pull() # generated in script 00_process

total_rev_cutoff <- total_rev_cutoff_2005 / gdp_defl_2005 
indiv_rev_cutoff <- indiv_rev_cutoff_2005 / gdp_defl_2005

edgetype <- "connectivity"

```
<br>

```{r eval=TRUE}


# y=2019
# p = "AST"
# iopac = "Astoria"
# wc_state = "W"
for(y in years){
  for(iopac in myports){ # updated to iopac 03-10-21, was p
    cat(iopac, "\n")
    
    ### pull fish ticket data for given year, port group ###
    
    dat <- fishtix %>%
      #filter(crab_year == 2008 & IOPAC == "Bodega Bay") # this line is just for running through for loop manually
      filter(crab_year == y & IOPAC == iopac) 
    
    ##### Create igraph object ####
    
    ## using participation_network function *if write_out = true, set pcid_choose/year_choose to p/y for file naming!
    tmp_g <- participation_network_crabyear(
      dat, 
      filter = TRUE, filter_subgraph = FALSE,
      edge_type="connectivity", pcid_choose=iopac, state_choose = NA, year_choose=y, min_vessels = vessel_cutoff, min_contribution = contr_cutoff, min_rev = total_rev_cutoff, min_rev_indiv = indiv_rev_cutoff,
      write_out=TRUE, out_dir = outdir_ports) # changed write_out=FALSE 12-31-2024
    
    if(any(is.na(tmp_g))){
      message("\n\n WARNING: No fishing at ", iopac, " in crab year ", y, "! \n\n") # updated to iopac 03-10-21, was No fishing at ", p, "
    } else if(vcount(tmp_g) > 0){
      ## add in the full metier names as vertex attribute
      filtered_metier_info <- as.data.frame(x = V(tmp_g)$name)
      colnames(filtered_metier_info) <- "SPGRPN2" #"metier.abbrev"
      filtered_metier_info <- left_join(filtered_metier_info, node_names, by = "SPGRPN2")
      V(tmp_g)$common_name <- as.character(filtered_metier_info$species_group_2_label2)
      ## save vertex colors as attribute
      vertex_cols <- vertex_color(g = tmp_g)
      V(tmp_g)$colors <- vertex_cols
      ## save objects
      rds_name <- ifelse(rm_crab,  paste0("igraph_", iopac, "_", y, "_totalrev", total_rev_cutoff,"_indivrev", indiv_rev_cutoff,"_", edgetype,"_rmCrab","_",Sys.Date(),".rds"),  paste0("igraph_", iopac, "_", y, "_", total_rev_cutoff,"_", indiv_rev_cutoff,"_", edgetype,"_",Sys.Date(),".rds")) # updated to iopac 03-10-21, was "igraph_", p
      saveRDS(tmp_g, file = here::here(outdir_ports, rds_name))
      
    # Calculate node & network-level statistics
    ## using the `net_stats` function
    g_stats <- net_stats(tmp_g, y = y, stats="both") # ran first with this argument 12-29-2021
    g_nodestats <- g_stats[[1]]
    g_netstats <- g_stats[[2]]
      # g_stats <- net_stats(tmp_g, y = y, stats="node") # added 12-29-2021. ran second with this argument 12-29-2021
    #g_stats <- net_stats(tmp_g, y = 2008, stats="network")
    ## add on period (closure/open) and port group info
    g_netstats <- g_netstats %>%
      #mutate(period = rep("early", times=dim(g_stats)[1])) %>%
      mutate(pcgroup = rep(iopac, times=dim(g_netstats)[1])) # updated to iopac 03-10-21, was rep(p
    
    g_nodestats <- g_nodestats %>%
      mutate(pcgroup = rep(iopac, times=dim(g_nodestats)[1]))
    } #end 'else' close-g has > 0 vertex
    
    
    #### SAVE NETWORK STATS OUTPUT ####
    ## merge network stats from closure / open period
    # if(length(V(tmp_g)$name) == 0){
    #   tmp_stats <- g_stats
    # } else{
    #   tmp_stats <- rbind(g_stats, late_stats)
    # }
    ## append to network stats output data frame
    if(exists('mystats3')){
      mystats3 <- rbind(mystats3, g_netstats)
    } else{
      mystats3 <- g_netstats
    }
    if(exists('mystats4')){
      mystats4 <- rbind(mystats4, g_nodestats)
    } else{
      mystats4 <- g_nodestats
    }
  } #end (p in myports)
  cat("\n finished with all port groups in crab year ", y, "\n-----\n")
} #end (y in years)
```
<br>

Write stats out to file. 2008_2015_CA_CrabPorts_NoSubgraph_01contribution.csv

remove duplicates, write out csv without duplicates (added 01-07-2025)
```{r eval=TRUE}
# ran first for network level stats 12-29-2021
filename <- paste0("VesselLevelNetworkStats_", min(years), "_", max(years), "_", length(myports), "pcgroups_", "totalrev", total_rev_cutoff,"_indivrev", indiv_rev_cutoff,"_", edgetype, "_", ifelse(rm_crab, 'rmCrab_','_'), contr_cutoff*100, "pContribution_",Sys.Date(),".csv")

# Note: 01-07-2025 need to add check to remove duplicated "WA Coast" rows. 

write.csv(mystats3 %>% distinct(),here::here(statdir, filename), row.names=FALSE)

# ran second for node level stats 12-29-2021
filename <- paste0("VesselLevelNodeStats_", min(years), "_", max(years), "_", length(myports), "pcgroups_", "totalrev", total_rev_cutoff,"_indivrev", indiv_rev_cutoff,"_", edgetype, "_", ifelse(rm_crab, 'rmCrab_','_'), contr_cutoff*100, "pContribution_",Sys.Date(),".csv")
# 
write.csv(mystats4 %>% distinct(),here::here(statdir, filename), row.names=FALSE)
```
<br>



---

Script Runtime:
```{r}
Sys.time() - script_start_time
```
