==> scripts/ESR 2026/02_plot_annual_networks_cciea_ESR2026.Rmd <==
---
title: "Plot Annual Participation Networks for CCIEA Ecosystem Status Report to the Pacific Fisheries Management Council"
author: "M. Fisher, J.F. Samhouri"
date: "Written Jan. 6, 2021. Last Run `r Sys.Date()`"
output:
  html_document:
    df_print: paged
    toc: yes
    toc_depth: '3'
  pdf_document:
    highlight: haddock
    number_sections: yes
    toc: yes
    toc_depth: '3'
geometry: margin=1in
subtitle: Preparation for network analysis in CCIEA ESR 2025
fontsize: 11pt
---

# Description

Plot annual fisheries participation networks (from igraph object) by port group, crab year.

Section 2 will generate network graphs that are colored according to the gear group used in each fishery. You have the option to use either the force-directed layout algorithm developed by [Fruchterman & Reingold (1991)](http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.13.8444&rep=rep1&type=pdf) or a simple circular layout. The Fruchterman & Reingold layout is informative but can result in overlap / odd clustering, especially if there is a fishery that has no / weakly connecting edges to other fisheries in the network.

Section 3 will generate simplified network graphs using a circular layout that highlight the Dungeness crab node (the visualizations used for Figure 1 in Fisher et al. 2021, PNAS). The plotting function in this section of the script could be adjusted to more flexibly highlight other fisheries.
 
This code depends on the code "01_create_networks_cciea.Rmd", which makes network graphs and calculates network statistics.

Make sure to define most_recent_igraph so that the date is associated with the most recent igraph files


<br>
```{r "setup", include=FALSE}
if(!require("here")) {install.packages("here")}
library(here)
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = here::here())

## start time for full script
script_start_time <- Sys.time()
```
<br>

This script requires the following packages. 
```{r packages, message=FALSE, warning=FALSE}
if(!require("tidyverse")) {install.packages("tidyverse")}
if(!require("ggplot2")) {install.packages("ggplot2")}
if(!require("igraph")) {install.packages("igraph")}
if(!require("RColorBrewer")) {install.packages("RColorBrewer")}
if(!require("stringr")) {install.packages("stringr")}
if(!require("magick")) {install.packages("magick")}
if(!require("grid")) {install.packages("grid")}
if(!require("cowplot")) {install.packages("cowplot")}

```
<br>

And calls the following functions:
```{r}
source(here::here('R/color_network_nodes.R'))
#source(here::here('R/plot_network.R'))
#source(here::here('R/plot_simple_network.R'))
source(here::here('R/plot_network_comparable.R')) # note this function will plot networks with (1) edge thickness according to edge weight, (2) node size according to the vertex attribute "importance", and (3) node colors according to the vertex attribute "colors". File names will be constructed using node attributes 'p' and 'year'. This means that the function can be used to plot vessel-level or aggregate (bulk) networks; simply change the file_suffix to tell the two apart if needed, and change the node attributes accordingly (e.g., whether V(g)$p represents port group or state). 
```
<br>


# User Inputs 

Select your directories.
```{r}
## location of igraph objects
indir_ports = 'data/networks/participation_vessel_ports'

## output directory for network viz
pngdir_ports = 'data/networks/participation_vessel_ports/plots/comparable'

processdir <- "data/input"
```
<br>

Specify file name for inflation adjustment factors.
```{r get_AFIvalues}

gdp_defl <- read.csv(file = paste(here(processdir), "inflation_adjustment.csv", sep = "/")) # generated in script 00_process

```


Identify the crab years and port groups that you would like to produce networks for. 
```{r}
## (crab) years
years <- seq(2004,2025)
# years <- 2008 # testing script, 1/6
# years <- 2018 # testing script, 1/8
# years <- 2014 # testing script, 1/27

## port groups
myports <- c("Puget Sound","North WA Coast","WA Coast","Other Coastal Washington","Astoria","Tillamook","Columbia River","Newport","Coos Bay","Brookings","Crescent City","Eureka","Morro Bay","Fort Bragg","Bodega Bay","San Francisco","San Diego","Monterey","Santa Barbara","Los Angeles","Unknown Ports" )
#myports <- c("BDG") # testing script, 1/6
#myports <- c("Columbia River") # testing script for single node networks, 1/7
# myports <- c("Monterey") # testing script, 1/8
# myports <- c("Bodega Bay") # testing script, 1/27/21
```
<br>

Do you want to plot using the Fruchterman & Reingold layout ["fr"], a circular layout ["c"], or both ["both"]?
```{r}
my_layout <- "both"
```
<br>


# 1: Read in data

## 1) Vessel-level networks summarized by port group

B)(ii) and C)(i) A minimum annual total fisheries revenue cutoff for each vessel of `$5,000` (USD in 2005; following Holland et al. diversification indices), and a minimum annual individual fisheries revenue cutoff for each vessel of `$500` (USD in 2005; following Holland et al. diversification indices). Edge type is connectivity.

```{r}

total_rev_cutoff_2005 <- 5000 #USD in 2005; following Holland et al. diversification indices
indiv_rev_cutoff_2005 <- 500 #USD in 2005; following Holland et al. diversification indices

gdp_defl_2005 <- gdp_defl %>% filter ( year == 2005) %>% pull() # generated in script 00_process

total_rev_cutoff <- total_rev_cutoff_2005 / gdp_defl_2005 
indiv_rev_cutoff <- indiv_rev_cutoff_2005 / gdp_defl_2005



edgetype <- "connectivity"
## To avoid inclusion of rare or minimal fishing activity, a given fishery must contribute to at least contr_cutoff% of a port's seasonal revenue to be included in the network data. This cutoff value can be increased if more stringent fishery inclusion criteria are preferred, and decreased (i.e. for summaries over shorter temporal / smaller spatial scales).
contr_cutoff <- 0.1

```
<br>

Read the igraph objects into a list. This for loop will also apply the functions to recolor and rename vertices
```{r}
#palette(brewer.pal(n = 10, name = "Set3"))
#pal3 <- brewer.pal(n = 10, name = "Set3")

most_recent_igraph <- '2026-01-15' # set date associated with igraph files

graphs_list <- list()
graph_names <- c()
i=1
for(iopac in myports){
  for(y in years){
  #y=2013 # testing script for single node networks, 1/7
  # y=2014 # testing script for single node networks, 1/27
    if(file.exists(here::here(indir_ports,paste0("igraph_", iopac, "_", y, "_totalrev", total_rev_cutoff,"_indivrev", indiv_rev_cutoff,"_", edgetype,"_rmCrab_",most_recent_igraph,".rds")))){
      tmpgraph <- readRDS(here::here(indir_ports,paste0("igraph_", iopac, "_", y, "_totalrev", total_rev_cutoff,"_indivrev", indiv_rev_cutoff,"_", edgetype,"_rmCrab_",most_recent_igraph,".rds")))
      V(tmpgraph)$p <- iopac
      V(tmpgraph)$year <- y
      V(tmpgraph)$colors <- vertex_color(tmpgraph)
      graphs_list[[i]] <- tmpgraph
      graph_names[i] <- paste0(iopac,"_",y)
      i = i + 1
      } else{ y = y +1 }
  }
}
names(graphs_list) <- graph_names
```
<br>


# 2: Plot with Color-Adjusted Nodes

Plots will be written out to .png files in the function. This is because the igraph `plot` function won't save into an object that can be returned.

```{r}
#lapply(graphs_list,plot_network, my_layout, outdir=pngdir)
plot_comparable_networks(graphs_list,outdir=pngdir_ports,color_labels=FALSE)
```
<br>


Put the 2025, 2024, and all Astoria and all Morro Bay circular graphs in their own directories
```{r}

list.of.files_2025 <- list.files(here::here(pngdir_ports), "2025_circular_compare")
file.copy(here::here(pngdir_ports,list.of.files_2025), here::here(pngdir_ports, "2025 only"), overwrite = TRUE)

list.of.files_2024 <- list.files(here::here(pngdir_ports), "2024_circular_compare")
file.copy(here::here(pngdir_ports,list.of.files_2024), here::here(pngdir_ports, "2024 only"), overwrite = TRUE)

list.of.files_astoria <- intersect(list.files(here::here(pngdir_ports), pattern = "Astoria"), list.files(here::here(pngdir_ports), pattern = "circular")) 
file.copy(here::here(pngdir_ports,list.of.files_astoria), here::here(pngdir_ports, "astoria"), overwrite = TRUE)

list.of.files_morrobay <- intersect(list.files(here::here(pngdir_ports), pattern = "Morro Bay"), list.files(here::here(pngdir_ports), pattern = "circular")) 
file.copy(here::here(pngdir_ports,list.of.files_morrobay), here::here(pngdir_ports, "morrobay"), overwrite = TRUE)


```
<br>

Assemble the 2022 vs 2025 networks for 2 California ports that historically had heavy salmon participation before the last 3 years of salmon fishery closures.
```{r}

bodega_2022 <- intersect(intersect(list.files(here::here(pngdir_ports), pattern = "Bodega"), list.files(here::here(pngdir_ports), pattern = "circular")), list.files(here::here(pngdir_ports), pattern = "2022")) 
bodega_2025 <- intersect(intersect(list.files(here::here(pngdir_ports), pattern = "Bodega"), list.files(here::here(pngdir_ports), pattern = "circular")), list.files(here::here(pngdir_ports), pattern = "2025")) 

img_bb2022 <- magick::image_read(
  here::here(pngdir_ports,
             bodega_2022)
)

#Get the actual dimensions (width and height in pixels)
info_bb2022 <- magick::image_info(img_bb2022)

img_bb2025 <- magick::image_read(
  here::here(pngdir_ports,
             bodega_2025)
)

#Get the actual dimensions (width and height in pixels)
info_bb2025 <- magick::image_info(img_bb2025)

#gg_bb2022 <-image_ggplot(img_bb2022) + labs(title = "Bodega Bay 2022")
gg_bb2022 <- ggplot() +
  annotation_raster(img_bb2022, xmin = 0, xmax = info_bb2022$width, ymin = 0, ymax = info_bb2022$height) +
  coord_fixed() +
  scale_x_continuous(limits = c(0, info_bb2022$width[1]), expand = c(0,0)) +
  scale_y_continuous(limits = c(0, info_bb2022$height[1]), expand = c(0,0)) +
  theme_void() + 
  labs(title = "Bodega Bay 2022")

#gg_bb2025 <-image_ggplot(img_bb2025) + labs(title = "Bodega Bay 2025")
gg_bb2025 <- ggplot() +
  annotation_raster(img_bb2025, xmin = 0, xmax = info_bb2025$width, ymin = 0, ymax = info_bb2025$height) +
  coord_fixed() +
  scale_x_continuous(limits = c(0, info_bb2025$width[1]), expand = c(0,0)) +
  scale_y_continuous(limits = c(0, info_bb2025$height[1]), expand = c(0,0)) +
  theme_void() + 
  labs(title = "Bodega Bay 2025")

bb2022_2025 <- plot_grid(gg_bb2022, gg_bb2025)

ggsave(here::here(pngdir_ports,
                paste0('bodega bay 2022 v 2025_',Sys.Date(),'.png')), bb2022_2025, width = 6, height = 10, dpi =600)


monterey_2022 <- intersect(intersect(list.files(here::here(pngdir_ports), pattern = "Monterey"), list.files(here::here(pngdir_ports), pattern = "circular")), list.files(here::here(pngdir_ports), pattern = "2022")) 
monterey_2025 <- intersect(intersect(list.files(here::here(pngdir_ports), pattern = "Monterey"), list.files(here::here(pngdir_ports), pattern = "circular")), list.files(here::here(pngdir_ports), pattern = "2025")) 

img_mt2022 <- magick::image_read(
  here::here(pngdir_ports,
             monterey_2022)
)

#Get the actual dimensions (width and height in pixels)
info_mt2022 <- magick::image_info(img_mt2022)

img_mt2025 <- magick::image_read(
  here::here(pngdir_ports,
             monterey_2025)
)

#Get the actual dimensions (width and height in pixels)
info_mt2025 <- magick::image_info(img_mt2025)

gg_mt2022 <- ggplot() +
  annotation_raster(img_mt2022, xmin = 0, xmax = info_mt2022$width, ymin = 0, ymax = info_mt2022$height) +
  coord_fixed() +
  scale_x_continuous(limits = c(0, info_mt2022$width[1]), expand = c(0,0)) +
  scale_y_continuous(limits = c(0, info_mt2022$height[1]), expand = c(0,0)) +
  theme_void() + 
  labs(title = "Monterey 2022")

#gg_mt2025 <-image_ggplot(img_mt2025) + labs(title = "Monterey 2025")
gg_mt2025 <- ggplot() +
  annotation_raster(img_mt2025, xmin = 0, xmax = info_mt2025$width, ymin = 0, ymax = info_mt2025$height) +
  coord_fixed() +
  scale_x_continuous(limits = c(0, info_mt2025$width[1]), expand = c(0,0)) +
  scale_y_continuous(limits = c(0, info_mt2025$height[1]), expand = c(0,0)) +
  theme_void() + 
  labs(title = "Monterey 2025")

mt2022_2025 <- plot_grid(gg_mt2022, gg_mt2025)

ggsave(here::here(pngdir_ports,
                paste0('monterey 2022 v 2025_',Sys.Date(),'.png')), mt2022_2025, width = 6, height = 10, dpi =600)




```
