==> scripts/ESR 2026/04_IOPAC-network-map-script-cciea-MARKDOWN_ESR2026.Rmd <==
---
title: "IOPAC-network-map-script-cciea-MARKDOWN"
author: "Amanda Phillips, with modifications by J.F. Samhouri"
date: "Written Jan. 19, 2022. Last Run `r Sys.Date()`"
output: html_document
subtitle: Preparation for network analysis in CCIEA ESR 2026
---

Clear the environment
```{r}
rm(list=ls())

# output to the ESR repo automatically
# esr_dir = paste0('~/GitHub/Ecosystem-Status-Report-FY2026/Figures/')  # nick
esr_dir = paste0('/Users/jameal.samhouri/Documents/Github repos/Ecosystem-Status-Report-FY2026/Figures/') # jameal
# change to reflect directory that ESR repo is in
esr_output = TRUE # set to false if not writing ESR repo

```


```{r setup, include=FALSE}
library(tidyverse)
library(ggmap)
library(maps)
library(ggthemes)
library(ggrepel)
library(ggimage)
library(cowplot)
library(here)

## output directory for network viz
pngdir_ports = 'data/networks/participation_vessel_ports/plots/comparable/2025 only/.'
fig_path = here::here(pngdir_ports)

#load network node images

thisyear <- 2025

# add date for most recent graphs
most_recent_igraph <- "2026-01-20"

#fig_path
thisyearsESRimages <- list.files(here::here(pngdir_ports), paste0("2025_circular_compare_",most_recent_igraph))

wa_fig_path <- here::here('data/networks/participation_vessel_ports/plots/comparable/')

knitr::opts_chunk$set(fig.width=6, fig.height=10, fig.path= fig_path, dpi = 600,
                      results = 'asis', concordance=F, echo=F, cache = F, warning = F, error=F, message = F)


```

```{r}




#wa
img_nwc <- magick::image_read(
  here::here(pngdir_ports,
  thisyearsESRimages[grep("North WA Coast", thisyearsESRimages)])
  )
img_wc <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("(?<!North )WA Coast", thisyearsESRimages, perl = TRUE)])
)
img_ps <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("Puget Sound", thisyearsESRimages)])
)

#or
img_ast <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("Astoria", thisyearsESRimages)])
) 
img_tl <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("Tillamook", thisyearsESRimages)])
)
img_nwp <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("Newport", thisyearsESRimages)])
)
img_cb <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("Coos Bay", thisyearsESRimages)])
)
img_brk <-  magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("Brookings", thisyearsESRimages)])
)
img_cr <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("Columbia River", thisyearsESRimages)])
)

#nca
img_cc <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("Crescent City", thisyearsESRimages)])
)
img_eu <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("Eureka", thisyearsESRimages)])
)
img_fb <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("Fort Bragg", thisyearsESRimages)])
)
img_bb <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("Bodega Bay", thisyearsESRimages)])
)
img_sf <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("San Francisco", thisyearsESRimages)])
)
img_mt <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("Monterey", thisyearsESRimages)])
)

#sca
img_sd <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("San Diego", thisyearsESRimages)])
)
img_sb <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("Santa Barbara", thisyearsESRimages)])
)
img_la <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("Los Angeles", thisyearsESRimages)])
)
img_mb <- magick::image_read(
  here::here(pngdir_ports,
             thisyearsESRimages[grep("Morro Bay", thisyearsESRimages)])
)


#load spreadsheet with lat-long coordinates for boxes
box <- read_csv(here::here('data','input','node_box_ref_esr2026.csv'))

```


```{r}
##### coastwide figure #####

#identify ports of interest to include in the boxes
# poi <- c('North WA Coast', 'Astoria', 'Fort Bragg', 'Morro Bay')
poi <- c('Puget Sound', 'Newport', 'Crescent City', 'Eureka')


#function to create the grey base-map for maps
base_map_iopac <- function(state, xlim, ylim) {
  states <- map_data("state")
  west_coast <- subset(states, region %in% state)
  base_map <- ggplot(data = west_coast) +
    geom_polygon(aes(x = long, y = lat, group = group), fill = "gray90", alpha = 0.5, color = "gray85") +
    coord_map(xlim = xlim, ylim = ylim) +
    theme_map()
  return(base_map)
  }

# filter function for boxes to include
box_filter <- function(region) {
  box <- box %>% filter(map_version == region)
  return(box)
  }

# function to create iopac map with boxes
box_iopac <- function(basemap, box_filter, region) {
  iopac_rect <- basemap + 
  geom_point(data = subset(box, map_version %in% region), aes(x=PRIMARY_LONGITUDE, y=PRIMARY_LATITUDE, fill = STATE), colour="black", pch=21, size = 2) +
  geom_text_repel(aes(x=PRIMARY_LONGITUDE, y=PRIMARY_LATITUDE, label= IOPAC), data = subset(box, map_version %in% region), size = 2.15, # AG changed text size 010425
                  box.padding = 0, point.padding = 0.25, hjust = -.25) +
  geom_rect(data = box_filter, aes(xmax = xmax, ymax = ymax, xmin = xmin, ymin=ymin), fill=NA, colour="darkgray") +
  geom_segment(data = subset(box_filter, segment_line == 'upper_right'), aes(x=PRIMARY_LONGITUDE, y=PRIMARY_LATITUDE, xend = xmax, yend = ymax), color = 'darkgray') +
  geom_segment(data = subset(box_filter, segment_line == 'lower_right'), aes(x=PRIMARY_LONGITUDE, y=PRIMARY_LATITUDE, xend = xmax, yend = ymin), color = 'darkgray') +
  geom_segment(data = subset(box_filter, segment_line == 'upper_left'), aes(x=PRIMARY_LONGITUDE, y=PRIMARY_LATITUDE, xend = xmin, yend = ymax), color = 'darkgray') +
  geom_segment(data = subset(box_filter, segment_line == 'lower_left'), aes(x=PRIMARY_LONGITUDE, y=PRIMARY_LATITUDE, xend = xmin, yend = ymin), color = 'darkgray') +
  theme(axis.title.x = element_blank(), axis.title.y = element_blank(), legend.position = "none") +
  scale_fill_manual(values = c('gray', 'black', 'white'))
  return(iopac_rect)
}


```



```{r iopac-node_full-coast_rect_font_m, fig.width=7, fig.height=9}

# AG changed image sizes below 010425

CW_map <- ggdraw(box_iopac(base_map_iopac(c("washington", "oregon", "california"), c(-130, -117), c(NA, NA)), box_filter('CW'),c('WA', 'OR', 'NCA', 'SCA'))) + # AG changed longitude limits 010425
  draw_image(img_ps, x = .06, y = .66, hjust = 0, vjust = 0, width = 0.31, height = 0.31) +
  draw_image(img_nwp, x = .08, y = .43, hjust = 0, vjust = 0, width = 0.31, height = 0.31) +
  draw_image(img_cc, x = .08, y = .18, hjust = 0, vjust = 0, width = 0.31, height = 0.31) +
  draw_image(img_eu, x = .27, y = .29, hjust = 0, vjust = 1, width = 0.31, height = 0.31) 

CW_map

ggsave(here::here(pngdir_ports,
                paste0('iopac-node_full-coast_rect_font_',thisyear,'.png')), CW_map, width = 6, height = 10, dpi =600)

# output to ESR repo
if(esr_output==TRUE){ggsave( paste0(esr_dir,'fpn-coast-',thisyear,'.png'), CW_map, width = 6, height = 10, dpi =600)}

```


```{r iopac-node_WA_rect_m, fig.width=6, fig.height=6}
WA_map <- ggdraw(box_iopac(base_map_iopac("washington", c(-130, -117), c(NA, NA)), 
                             box_filter('WA'), 
                             'WA')) + 
  draw_image(img_ps, x = .39, y = .66, hjust = 0, vjust = 0, width = 0.3, height = 0.26) +
  draw_image(img_nwc, x = .14, y = .34, hjust = 0, vjust = 0, width = 0.3, height = 0.26) +
  draw_image(img_wc, x = .39, y = .05, hjust = 0, vjust = 0, width = 0.3, height = 0.26) 
  
WA_map

ggsave(here::here(pngdir_ports,
                  paste0('iopac-node_WA_rect_',thisyear,'.png')
                  ), WA_map, width = 6, height = 6, dpi =600)

if(esr_output==TRUE){ggsave( paste0(esr_dir,'fpn-wa-',thisyear,'.png'), WA_map, width = 6, height = 6, dpi =600)}
```


```{r iopac-node_OR_rect_m, fig.width=6, fig.height=6}

OR_map <- ggdraw(box_iopac(base_map_iopac("oregon", c(-130, -117), c(NA, NA)), 
                             box_filter('OR'), 
                             'OR')) + 
  draw_image(img_cr, x = .61, y = .65, hjust = 0, vjust = 0, width = 0.29, height = 0.29) +
  draw_image(img_ast, x = .35, y = .695, hjust = 0, vjust = 0, width = 0.29, height = 0.25) +
  draw_image(img_tl, x = .04, y = .61, hjust = 0, vjust = 0, width = 0.29, height = 0.29) +
  draw_image(img_nwp, x = .12, y = .34, hjust = 0, vjust = 0, width = 0.28, height = 0.28) +
  draw_image(img_cb, x = .11, y = .08, hjust = 0, vjust = 0, width = 0.26, height = 0.26) +
  draw_image(img_brk, x = .515, y = .04, hjust = 0, vjust = 0, width = 0.28, height = 0.28)

OR_map

ggsave(here::here(pngdir_ports,
                  paste0('iopac-node_OR_rect_',thisyear,'.png')
                  ), OR_map, width = 6, height = 6, dpi =600)

if(esr_output==TRUE){ggsave( paste0(esr_dir,'fpn-or-',thisyear,'.png'), OR_map, width = 6, height = 6, dpi =600)}
```

```{r iopac-node_NCA_rect_m, fig.width=7, fig.height=6}

NCA_map <- ggdraw(box_iopac(base_map_iopac("california", c(-130, -117), c(NA, NA)), 
                            box_filter('NCA'), 
                            'NCA')) + 
  draw_image(img_cc, x = .245, y = .71, hjust = 0, vjust = 0, width = 0.23, height = 0.23) +
  draw_image(img_eu, x = .505, y = .72, hjust = 0, vjust = 0, width = 0.24, height = 0.24) +
  draw_image(img_fb, x = .200, y = .49, hjust = 0, vjust = 0, width = 0.22, height = 0.22) +
  draw_image(img_bb, x = .625, y = .48, hjust = 0, vjust = 0, width = 0.22, height = 0.24) +
  draw_image(img_sf, x = .225, y = .25, hjust = 0, vjust = 0, width = 0.24, height = 0.24) +
  draw_image(img_mt, x = .615, y = .26, hjust = 0, vjust = 0, width = 0.24, height = 0.24)

NCA_map

ggsave(here::here(pngdir_ports,
                  paste0('iopac-node_NCA_rect_',thisyear,'.png')
                  ), NCA_map, width = 6, height = 6, dpi =600)
                  
                  
if(esr_output==TRUE){ggsave( paste0(esr_dir,'fpn-nca-',thisyear,'.png'), NCA_map, width = 6, height = 6, dpi =600)}
```


```{r iopac-node_SCA_rect_m, fig.width=6, fig.height=5}

SCA_map <- ggdraw(box_iopac(base_map_iopac("california", c(-130, -117), c(NA, NA)), 
                            box_filter('SCA'), 
                            'SCA')) + 
  draw_image(img_mb, x = .25, y = .325, hjust = 0, vjust = 0, width = 0.19, height = 0.20) +
  draw_image(img_sb, x = .46, y = .26, hjust = 0, vjust = 0, width = 0.20, height = 0.21) +
  draw_image(img_la, x = .43, y = .04, hjust = 0, vjust = 0, width = 0.20, height = 0.21) +
  draw_image(img_sd, x = .65, y = .04, hjust = 0, vjust = 0, width = 0.20, height = 0.21) 

SCA_map

ggsave(here::here(pngdir_ports,
                  paste0('iopac-node_SCA_rect_',thisyear,'.png')
                  ), SCA_map, width = 6, height = 6, dpi =600)

if(esr_output==TRUE){ggsave( paste0(esr_dir,'fpn-sca-',thisyear,'.png'), SCA_map, width = 6, height = 6, dpi =600)}
```


```{r All Fisheries Participation Networks 2024-2025, by IOPAC Port Group, fig.width = 16, fig.height = 16}
# png(here::here(pngdir_ports, paste0("All Fisheries Participation Networks 2020-2021, by IOPAC Port Group.png")),
#     res=200, height=1600,width=1600)  # ,height=1400,width=1600

p_all <- ggplot(data.frame(x = 1:10, y = 1:10), aes(x, y)) +
  geom_point(size = 1, colour=NA) +
  draw_image(img_nwc , x = 1, y = 9, scale = 2) +
  draw_image(img_wc , x = 3, y = 9, scale = 2) +
  draw_image(img_ps , x = 5, y = 9, scale = 2) +
  draw_image(img_ast , x = 7, y = 9, scale = 2) +
  draw_image(img_tl , x = 9, y = 9, scale = 2) +
  draw_image(img_cr , x = 1, y = 7, scale = 2) +
  draw_image(img_nwp , x = 3, y = 7, scale = 2) +
  draw_image(img_brk , x = 5, y = 7, scale = 2) +
  draw_image(img_cb , x = 7, y = 7, scale = 2) +
  draw_image(img_cc , x = 9, y = 7, scale = 2) +
  draw_image(img_eu , x = 1, y = 5, scale = 2) +
  draw_image(img_fb , x = 3, y = 5, scale = 2) +
  draw_image(img_bb , x = 5, y = 5, scale = 2) +
  draw_image(img_sf , x = 7, y = 5, scale = 2) +
  draw_image(img_mt , x = 9, y = 5, scale = 2) +
  draw_image(img_mb , x = 1, y = 3, scale = 2) +
  draw_image(img_sb, x = 3, y = 3, scale = 2) + 
  draw_image(img_la , x = 5, y = 3, scale = 2) +
  draw_image(img_sd , x = 7, y = 3, scale = 2) + 
  theme_void()

p_all

ggsave(here::here(pngdir_ports,
                  paste0('All Fisheries Participation Networks ', thisyear, ' by IOPAC Port Group.png')
                  ), p_all, width = 6, height = 6, dpi =600)

```


<!-- ```{r WA Coast Fisheries Participation Networks 2016-2021, fig.width = 16, fig.height = 16} -->

<!-- wa_fig_path <- here::here('data/networks/participation_vessel_ports/plots/comparable') -->
<!-- wa_ESRimages <- list.files(wa_fig_path, "^WA Coast") -->

<!-- img_wc_2020 <- img_wc -->

<!-- img_wc_2019 <- magick::image_read( -->
<!--   here::here(wa_fig_path, -->
<!--              wa_ESRimages[grep("^WA Coast_2019", wa_ESRimages)]) -->
<!-- ) -->

<!-- img_wc_2018 <- magick::image_read( -->
<!--   here::here(wa_fig_path, -->
<!--              wa_ESRimages[grep("^WA Coast_2018", wa_ESRimages)]) -->
<!-- ) -->

<!-- img_wc_2017 <- magick::image_read( -->
<!--   here::here(wa_fig_path, -->
<!--              wa_ESRimages[grep("^WA Coast_2017", wa_ESRimages)]) -->
<!-- ) -->

<!-- img_wc_2016 <- magick::image_read( -->
<!--   here::here(wa_fig_path, -->
<!--              wa_ESRimages[grep("^WA Coast_2016", wa_ESRimages)]) -->
<!-- ) -->

<!-- img_wc_2015 <- magick::image_read( -->
<!--   here::here(wa_fig_path, -->
<!--              wa_ESRimages[grep("^WA Coast_2015", wa_ESRimages)]) -->
<!-- ) -->

<!-- # png(here::here(pngdir_ports, paste0("All Fisheries Participation Networks 2020-2021, by IOPAC Port Group.png")), -->
<!-- #     res=200, height=1600,width=1600)  # ,height=1400,width=1600 -->

<!-- # ggplot(data.frame(x = 1:6, y = 1:6), aes(x, y)) + -->
<!-- #   geom_point(size = 1, colour=NA) + -->
<!-- #   draw_image(img_wc_2015 , x = 1, y = 4, scale = 2) + -->
<!-- #   draw_image(img_wc_2016 , x = 3, y = 4, scale = 2) + -->
<!-- #   draw_image(img_wc_2017 , x = 5, y = 4, scale = 2) + -->
<!-- #   draw_image(img_wc_2018 , x = 1, y = 2, scale = 2) + -->
<!-- #   draw_image(img_wc_2019 , x = 3, y = 2, scale = 2) + -->
<!-- #   draw_image(img_wc_2020 , x = 5, y = 2, scale = 2) + -->
<!-- #   theme_void() -->

<!-- # ggplot(data.frame(x = 1:6, y = 1:6), aes(x, y)) + -->
<!-- #   geom_point(size = 1, colour=NA) + -->
<!-- #   draw_image(img_wc_2015 , x = 2, y = 5, scale = 2) + -->
<!-- #   draw_image(img_wc_2016 , x = 4, y = 5, scale = 2) + -->
<!-- #   draw_image(img_wc_2017 , x = 2, y = 3, scale = 2) + -->
<!-- #   draw_image(img_wc_2018 , x = 4, y = 3, scale = 2) + -->
<!-- #   draw_image(img_wc_2019 , x = 2, y = 1, scale = 2) + -->
<!-- #   draw_image(img_wc_2020 , x = 4, y = 1, scale = 2) + -->
<!-- #   theme_void() -->

<!-- p_wc_2015 <- ggplot(data.frame(x = 1:2, y = 1:2), aes(x, y)) + -->
<!--   geom_point(size = 1, colour=NA) + -->
<!--   draw_image(img_wc_2015 , x = 1, y = 1, scale = 1.5) + -->
<!--   #labs(title = '2016') + -->
<!--   theme_void() #+ -->
<!--   # theme( -->
<!--   #   plot.title.position = "plot", -->
<!--   #   plot.title = element_text(size = 18, hjust = 0, vjust = 0.5,  -->
<!--   #                               margin = margin(b = 0.2, unit = "cm")) -->
<!--   # ) -->

<!-- p_wc_2016 <- ggplot(data.frame(x = 1:2, y = 1:2), aes(x, y)) + -->
<!--   geom_point(size = 1, colour=NA) + -->
<!--   draw_image(img_wc_2016 , x = 1, y = 1, scale = 1.5) + -->
<!--   #labs(title = '2017') + -->
<!--   theme_void() #+ -->
<!--   # theme( -->
<!--   #   plot.title.position = "plot", -->
<!--   #   plot.title = element_text(size = 18, hjust = 0, vjust = 0.5,  -->
<!--   #                               margin = margin(b = 0.2, unit = "cm")) -->
<!--   # ) -->

<!-- p_wc_2017 <- ggplot(data.frame(x = 1:2, y = 1:2), aes(x, y)) + -->
<!--   geom_point(size = 1, colour=NA) + -->
<!--   draw_image(img_wc_2017 , x = 1, y = 1, scale = 1.5) + -->
<!--   theme_void() -->

<!-- p_wc_2018 <- ggplot(data.frame(x = 1:2, y = 1:2), aes(x, y)) + -->
<!--   geom_point(size = 1, colour=NA) + -->
<!--   draw_image(img_wc_2018 , x = 1, y = 1, scale = 1.5) + -->
<!--   theme_void() -->

<!-- p_wc_2019 <- ggplot(data.frame(x = 1:2, y = 1:2), aes(x, y)) + -->
<!--   geom_point(size = 1, colour=NA) + -->
<!--   draw_image(img_wc_2019 , x = 1, y = 1, scale = 1.5) + -->
<!--   theme_void() -->

<!-- p_wc_2020 <- ggplot(data.frame(x = 1:2, y = 1:2), aes(x, y)) + -->
<!--   geom_point(size = 1, colour=NA) + -->
<!--   draw_image(img_wc_2020 , x = 1, y = 1, scale = 1.5) + -->
<!--   theme_void() -->

<!-- p_wa_2015_2020 <- plot_grid(p_wc_2015, p_wc_2016, p_wc_2017, p_wc_2018, p_wc_2019, p_wc_2020, -->
<!--           labels = c('2016', '2017', '2018', '2019', '2020', '2021'),  -->
<!--           label_size = 12, label_y = c(1, 1, 1.1, 1.1, 1.1, 1.1), align = "hv", rel_widths = c(1, 1, 1, 1, 1, 1), ncol = 2 -->
<!-- ) #label_x = 0.4, c(0.35, 0.35, 0.35, 0.35, 0.35, 0.35) -->

<!-- p_wa_2015_2020 -->

<!-- ggsave(here::here(pngdir_ports,'WA Coast Fisheries Participation Networks 2016-2021.png'),  -->
<!--        p_wa_2015_2020,  -->
<!--        width = 7, height = 5, dpi =600) -->

<!-- ``` -->
<!-- <br> -->

<!-- ```{r Coos Bay Fisheries Participation Networks 2016-2021, fig.width = 16, fig.height = 16} -->

<!-- wa_fig_path <- here::here('data/networks/participation_vessel_ports/plots/comparable') -->
<!-- cb_ESRimages <- list.files(wa_fig_path, "Coos Bay") -->

<!-- img_cb_2020 <- img_cb -->

<!-- img_cb_2019 <- magick::image_read( -->
<!--   here::here(wa_fig_path, -->
<!--              cb_ESRimages[grep("Coos Bay_2019", cb_ESRimages)]) -->
<!-- ) -->

<!-- img_cb_2018 <- magick::image_read( -->
<!--   here::here(wa_fig_path, -->
<!--              cb_ESRimages[grep("Coos Bay_2018", cb_ESRimages)]) -->
<!-- ) -->

<!-- img_cb_2017 <- magick::image_read( -->
<!--   here::here(wa_fig_path, -->
<!--              cb_ESRimages[grep("Coos Bay_2017", cb_ESRimages)]) -->
<!-- ) -->

<!-- img_cb_2016 <- magick::image_read( -->
<!--   here::here(wa_fig_path, -->
<!--              cb_ESRimages[grep("Coos Bay_2016", cb_ESRimages)]) -->
<!-- ) -->

<!-- img_cb_2015 <- magick::image_read( -->
<!--   here::here(wa_fig_path, -->
<!--              cb_ESRimages[grep("Coos Bay_2015", cb_ESRimages)]) -->
<!-- ) -->

<!-- p_cb_2015 <- ggplot(data.frame(x = 1:2, y = 1:2), aes(x, y)) + -->
<!--   geom_point(size = 1, colour=NA) + -->
<!--   draw_image(img_cb_2015 , x = 1, y = 1, scale = 1.5) + -->
<!--   theme_void()  -->

<!-- p_cb_2016 <- ggplot(data.frame(x = 1:2, y = 1:2), aes(x, y)) + -->
<!--   geom_point(size = 1, colour=NA) + -->
<!--   draw_image(img_cb_2016 , x = 1, y = 1, scale = 1.5) + -->
<!--   theme_void()  -->

<!-- p_cb_2017 <- ggplot(data.frame(x = 1:2, y = 1:2), aes(x, y)) + -->
<!--   geom_point(size = 1, colour=NA) + -->
<!--   draw_image(img_cb_2017 , x = 1, y = 1, scale = 1.5) + -->
<!--   theme_void() -->

<!-- p_cb_2018 <- ggplot(data.frame(x = 1:2, y = 1:2), aes(x, y)) + -->
<!--   geom_point(size = 1, colour=NA) + -->
<!--   draw_image(img_cb_2018 , x = 1, y = 1, scale = 1.5) + -->
<!--   theme_void() -->

<!-- p_cb_2019 <- ggplot(data.frame(x = 1:2, y = 1:2), aes(x, y)) + -->
<!--   geom_point(size = 1, colour=NA) + -->
<!--   draw_image(img_cb_2019 , x = 1, y = 1, scale = 1.5) + -->
<!--   theme_void() -->

<!-- p_cb_2020 <- ggplot(data.frame(x = 1:2, y = 1:2), aes(x, y)) + -->
<!--   geom_point(size = 1, colour=NA) + -->
<!--   draw_image(img_cb_2020 , x = 1, y = 1, scale = 1.5) + -->
<!--   theme_void() -->

<!-- p_cb_2015_2020 <- plot_grid(p_cb_2015, p_cb_2016, p_cb_2017, p_cb_2018, p_cb_2019, p_cb_2020, -->
<!--           labels = c('2016', '2017', '2018', '2019', '2020', '2021'),  -->
<!--           label_size = 12, label_y = c(1, 1, 1.1, 1.1, 1.1, 1.1), align = "hv", rel_widths = c(1, 1, 1, 1, 1, 1), ncol = 2 -->
<!-- ) #label_x = 0.4, c(0.35, 0.35, 0.35, 0.35, 0.35, 0.35) -->

<!-- p_cb_2015_2020 -->

<!-- ggsave(here::here(pngdir_ports,'Coos Bay Fisheries Participation Networks 2016-2021.png'),  -->
<!--        p_cb_2015_2020,  -->
<!--        width = 7, height = 5, dpi =600 -->
<!--        ) -->

<!-- ``` -->
<!-- <br> -->

<!-- ```{r Coos Bay Fisheries Participation Networks 2015 v 2021, fig.width = 16, fig.height = 16} -->

<!-- wa_fig_path <- here::here('data/networks/participation_vessel_ports/plots/comparable') -->
<!-- cb_ESRimages <- list.files(wa_fig_path, "Coos Bay") -->

<!-- # img_cb_2020 <- img_cb -->

<!-- img_cb_2012 <- magick::image_read( -->
<!--   here::here(wa_fig_path, -->
<!--              cb_ESRimages[grep("Coos Bay_2012", cb_ESRimages)]) -->
<!-- ) -->

<!-- p_cb_2012 <- ggplot(data.frame(x = 1:2, y = 1:2), aes(x, y)) + -->
<!--   geom_point(size = 1, colour=NA) + -->
<!--   draw_image(img_cb_2012 , x = 1, y = 1, scale = 1.1) + -->
<!--   theme_void()  -->

<!-- p_cb_2020_2 <- ggplot(data.frame(x = 1:2, y = 1:2), aes(x, y)) + -->
<!--   geom_point(size = 1, colour=NA) + -->
<!--   draw_image(img_cb_2020 , x = 1, y = 1, scale = 1.1) + -->
<!--   theme_void() -->

<!-- p_cb_2012_v_2020 <- plot_grid(p_cb_2012, p_cb_2020_2, -->
<!--           labels = c('2013', '2021'),  -->
<!--           label_size = 12, label_y = c(1, 1), align = "hv", rel_widths = c(1, 1), ncol = 2 -->
<!-- ) #label_x = 0.4, c(0.35, 0.35, 0.35, 0.35, 0.35, 0.35) -->

<!-- p_cb_2012_v_2020 -->

<!-- ggsave(here::here(pngdir_ports,'Coos Bay Fisheries Participation Networks 2013 and 2021.png'),  -->
<!--        p_cb_2012_v_2020,  -->
<!--        width = 7, height = 5, dpi =600 -->
<!--        ) -->

<!-- ``` -->
<!-- <br> -->
